FilterInteractions

Search in a GInteraction object which interactions correspond ti a target list and return a list of index or filter a matrices list according to target and a selection function.

Usage

FilterInteractions(
  matrices.lst = NULL,
  interarctions.gni = NULL,
  target.lst = NULL,
  selection.fun = function() {
     Reduce(intersect, interarctions.ndx_lst)
 }
)

Arguments

matrices.lst

<List[matrix]>: The matrices list to filter. If is not NULL, the function will return the filtred matrices list, else return a list of index.

interarctions.gni

<GInteractions>: The GInteraction object on which compute the filter.

target.lst

<List>: a nammed list that describe the target.

selection.fun

<function>: A function that described how the target must be cross. (Defaul intersection of all targets)

Value

A list of elements index or a filtred matrices list with attributes updates.

Examples

library(GenomicED)
data("submatrixPF_Ctrl.mtx_lst")
target.lst <- list(
  anchor.Beaf.name = c("Beaf32_8","Beaf32_15"),
  bait.Tss.name    = c("FBgn0031214","FBgn0005278"),
  name             = c("2L:74_2L:77"),
  distance         = function(columnElement){
    return(14000==columnElement || columnElement == 3000)
    }
)
# Extraction on InteractionSet
FilterInteractions(
  interarctions.gni = attributes(submatrixPF_Ctrl.mtx_lst)$interactions,
  target.lst        = target.lst,
  selection.fun     = NULL
)
#> $anchor.Beaf.name
#> [1] 1 3 4
#> 
#> $bait.Tss.name
#> [1] 1 2 3
#> 
#> $name
#> [1] 1
#> 
#> $distance
#> [1]  1  3  8 24 38 49 63 65 80
#> 

selection.fun = function(){
  Reduce(intersect, list(anchor.Beaf.name, bait.Tss.name ,distance) ) |>
  setdiff(name)
}

# Extraction on matrices list and with selection
FilterInteractions(
  matrices.lst = submatrixPF_Ctrl.mtx_lst,
  target.lst        = target.lst,
  selection.fun     = selection.fun
)
#> $`2L:107_2L:121`
#>           [,1]      [,2]      [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 0.000000 0.0000000 0.5297419 0.000000 1.108568 0.000000 0.000000 0.000000
#>  [2,] 0.000000 0.0000000 2.3123753 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [3,] 2.958131 0.0000000 0.0000000 2.139807 0.000000 1.286100 0.000000 0.000000
#>  [4,] 0.000000 0.0000000 0.0000000 1.466081 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.0000000 0.0000000 0.000000 0.000000 0.000000 2.045658 0.000000
#>  [6,] 0.000000 0.0000000 0.0000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [7,] 2.156886 0.0000000 0.0000000 0.000000 0.000000 0.000000 1.693100 2.782898
#>  [8,] 2.437283 0.8341782 0.0000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,] 1.878906 0.0000000 3.0109010 0.000000 5.779325 0.000000 1.959226 0.000000
#> [10,] 1.951472 3.4909725 0.0000000 4.173593 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 1.5685319 3.6822057 0.000000 0.000000 0.000000 0.000000 0.000000
#> [12,] 2.674218 0.0000000 0.0000000 3.479809 0.000000 2.087023 0.000000 2.178157
#> [13,] 0.000000 0.0000000 0.0000000 3.897979 3.738908 0.000000 0.000000 0.000000
#> [14,] 0.000000 1.0805069 2.0427310 0.000000 0.000000 0.000000 0.000000 0.000000
#> [15,] 0.000000 0.9250376 0.0000000 0.000000 0.000000 2.211734 0.000000 0.000000
#> [16,]       NA 0.0000000 0.0000000 1.598032 0.000000 0.000000 0.000000 0.000000
#> [17,]       NA        NA 0.0000000 0.000000 0.000000 2.082856 2.135369 0.000000
#> [18,]       NA        NA        NA 0.000000 0.000000 0.000000 0.000000 3.563908
#> [19,]       NA        NA        NA       NA 0.000000 2.344870 0.000000 0.000000
#> [20,]       NA        NA        NA       NA       NA 0.000000 0.000000 0.000000
#> [21,]       NA        NA        NA       NA       NA       NA 2.276137 0.000000
#> [22,]       NA        NA        NA       NA       NA       NA       NA 0.000000
#> [23,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [24,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [25,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [26,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [27,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [28,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [29,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [30,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [31,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [32,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [33,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [34,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [35,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [36,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [37,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [38,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [39,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [40,]       NA        NA        NA       NA       NA       NA       NA       NA
#> [41,]       NA        NA        NA       NA       NA       NA       NA       NA
#>            [,9]     [,10]     [,11]    [,12]    [,13]     [,14]     [,15]
#>  [1,] 0.0000000 0.6309677 0.0000000 0.000000 0.000000 0.0000000 0.8179082
#>  [2,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.0000000 2.2148798 0.000000 0.000000 0.0000000 0.0000000
#>  [4,] 1.0568750 0.0000000 0.0000000 0.000000 0.000000 1.2346566 0.0000000
#>  [5,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#>  [6,] 2.1355312 0.0000000 0.0000000 1.562632 0.000000 0.0000000 0.0000000
#>  [7,] 0.0000000 0.0000000 2.1186442 0.000000 0.000000 0.0000000 0.8296640
#>  [8,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.4825633 0.0000000
#>  [9,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> [10,] 0.0000000 0.0000000 0.0000000 0.000000 2.077487 2.1412535 0.0000000
#> [11,] 0.0000000 0.0000000 0.9881868 0.000000 0.000000 0.0000000 3.5791350
#> [12,] 0.3306442 0.0000000 2.2194939 0.000000 0.000000 0.0000000 0.0000000
#> [13,] 0.0000000 1.9560894 0.0000000 0.000000 3.229433 0.0000000 0.0000000
#> [14,] 0.0000000 1.1215968 0.0000000 0.000000 0.000000 0.0000000 1.2940578
#> [15,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> [16,] 1.3219687 0.0000000 0.6505961 1.734163 0.000000 0.4416105 0.0000000
#> [17,] 0.0000000 0.0000000 0.7581898 0.000000 0.000000 0.0000000 1.2983716
#> [18,] 0.9246346 2.0556383 0.0000000 1.996511 0.000000 0.0000000 0.0000000
#> [19,] 0.0000000 3.2716222 0.0000000 2.209554 0.000000 0.0000000 0.0000000
#> [20,] 0.0000000 0.0000000 0.0000000 1.860589 0.000000 0.0000000 0.0000000
#> [21,] 0.0000000 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> [22,] 0.5777135 0.0000000 1.8463304 1.569322 3.020721 1.6848464 0.0000000
#> [23,] 1.5169718 0.0000000 0.0000000 0.000000 0.000000 0.0000000 0.0000000
#> [24,]        NA 1.6781235 0.0000000 0.000000 3.185435 0.0000000 1.3574467
#> [25,]        NA        NA 0.7214578 0.000000 3.329250 1.5573801 0.0000000
#> [26,]        NA        NA        NA 0.000000 0.000000 0.0000000 0.0000000
#> [27,]        NA        NA        NA       NA 0.000000 0.0000000 0.0000000
#> [28,]        NA        NA        NA       NA       NA 1.8871852 0.0000000
#> [29,]        NA        NA        NA       NA       NA        NA 0.0000000
#> [30,]        NA        NA        NA       NA       NA        NA        NA
#> [31,]        NA        NA        NA       NA       NA        NA        NA
#> [32,]        NA        NA        NA       NA       NA        NA        NA
#> [33,]        NA        NA        NA       NA       NA        NA        NA
#> [34,]        NA        NA        NA       NA       NA        NA        NA
#> [35,]        NA        NA        NA       NA       NA        NA        NA
#> [36,]        NA        NA        NA       NA       NA        NA        NA
#> [37,]        NA        NA        NA       NA       NA        NA        NA
#> [38,]        NA        NA        NA       NA       NA        NA        NA
#> [39,]        NA        NA        NA       NA       NA        NA        NA
#> [40,]        NA        NA        NA       NA       NA        NA        NA
#> [41,]        NA        NA        NA       NA       NA        NA        NA
#>          [,16]     [,17]     [,18]    [,19]     [,20]     [,21]     [,22]
#>  [1,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.000000 0.0000000 0.0000000 1.525332 0.0000000 0.0000000 0.0000000
#>  [5,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.6048788 0.0000000
#>  [6,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [7,] 1.072001 0.0000000 2.5576630 0.000000 0.0000000 1.0404003 0.0000000
#>  [8,] 0.000000 1.1549203 0.0000000 1.208733 0.0000000 0.0000000 1.3595614
#>  [9,] 0.000000 0.0000000 0.5275351 0.000000 0.0000000 1.6110880 0.0000000
#> [10,] 0.000000 0.0000000 0.0000000 0.907349 0.0000000 1.2620573 0.0000000
#> [11,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.5119065 0.6231812
#> [12,] 1.516662 0.0000000 0.0000000 0.000000 0.4762146 0.0000000 0.6128533
#> [13,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> [14,] 0.000000 1.7248671 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> [15,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.4618533 0.0000000
#> [16,] 0.000000 0.0000000 0.5317324 0.000000 0.0000000 0.0000000 0.0000000
#> [17,] 0.000000 0.4379424 0.0000000 1.878967 0.0000000 0.0000000 0.0000000
#> [18,] 0.000000 0.0000000 1.5877159 1.890131 0.0000000 0.0000000 0.0000000
#> [19,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> [20,] 0.000000 1.2577319 0.0000000 0.000000 1.2296215 0.0000000 1.2168779
#> [21,] 0.000000 0.0000000 1.2413720 0.000000 0.0000000 2.0247781 0.0000000
#> [22,] 0.000000 0.0000000 0.0000000 1.845688 0.0000000 0.0000000 0.0000000
#> [23,] 0.000000 0.0000000 0.0000000 0.000000 2.3270350 1.4408004 0.0000000
#> [24,] 0.000000 2.1675289 1.4236422 1.978526 0.0000000 0.0000000 2.6371455
#> [25,] 0.000000 0.0000000 0.0000000 0.000000 1.2401872 0.0000000 0.0000000
#> [26,] 1.676673 3.0994139 0.0000000 0.000000 0.0000000 0.7546380 0.0000000
#> [27,] 0.000000 0.0000000 0.0000000 1.989344 0.0000000 0.0000000 1.1523938
#> [28,] 0.000000 0.0000000 1.0379076 0.000000 0.0000000 0.0000000 0.3076864
#> [29,] 0.000000 0.0000000 0.0000000 0.000000 0.0000000 2.1900650 0.5035289
#> [30,] 0.000000 1.4831592 0.0000000 0.000000 2.0331670 0.0000000 0.5802198
#> [31,]       NA 0.0000000 0.0000000 0.000000 1.0627156 0.0000000 0.0000000
#> [32,]       NA        NA 0.0000000 0.000000 1.5727552 0.0000000 1.9269785
#> [33,]       NA        NA        NA 0.000000 0.0000000 0.0000000 0.0000000
#> [34,]       NA        NA        NA       NA 0.0000000 0.0000000 2.2842069
#> [35,]       NA        NA        NA       NA        NA 0.0000000 1.1608269
#> [36,]       NA        NA        NA       NA        NA        NA 0.0000000
#> [37,]       NA        NA        NA       NA        NA        NA        NA
#> [38,]       NA        NA        NA       NA        NA        NA        NA
#> [39,]       NA        NA        NA       NA        NA        NA        NA
#> [40,]       NA        NA        NA       NA        NA        NA        NA
#> [41,]       NA        NA        NA       NA        NA        NA        NA
#>          [,23]     [,24]     [,25]     [,26]     [,27]     [,28]    [,29]
#>  [1,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [2,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [3,] 0.000000 0.0000000 1.2256362 0.8787679 0.0000000 0.0000000 1.804757
#>  [4,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.8191016 0.000000
#>  [5,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [6,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [7,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [8,] 0.000000 0.0000000 0.5590918 0.0000000 0.0000000 0.0000000 0.000000
#>  [9,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [10,] 0.000000 0.0000000 0.0000000 1.2254941 0.0000000 0.0000000 0.000000
#> [11,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [12,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [13,] 0.000000 0.0000000 2.2442172 0.0000000 0.0000000 0.0000000 0.000000
#> [14,] 0.000000 0.0000000 0.0000000 0.6380129 0.0000000 1.6257743 0.000000
#> [15,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [16,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [17,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [18,] 0.000000 0.0000000 0.0000000 1.0325822 0.0000000 0.0000000 0.000000
#> [19,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [20,] 0.000000 0.3107510 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [21,] 0.000000 0.6393787 0.0000000 1.5915717 0.0000000 0.0000000 0.000000
#> [22,] 0.000000 0.0000000 2.2655265 0.0000000 0.0000000 0.0000000 0.000000
#> [23,] 0.409167 0.0000000 1.1775584 0.5781391 0.0000000 0.0000000 0.000000
#> [24,] 1.806444 0.0000000 0.0000000 0.0000000 0.8695775 0.0000000 0.000000
#> [25,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [26,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [27,] 1.596562 1.8188438 0.0000000 0.0000000 0.0000000 1.7012625 0.000000
#> [28,] 0.000000 0.0000000 2.8199500 0.0000000 0.0000000 0.0000000 0.000000
#> [29,] 0.000000 0.0000000 0.9319323 0.0000000 0.0000000 0.0000000 0.000000
#> [30,] 0.000000 0.0000000 0.0000000 0.0000000 0.4149599 0.0000000 0.000000
#> [31,] 0.000000 1.4643464 0.0000000 0.0000000 0.0000000 1.3108084 0.000000
#> [32,] 0.000000 0.0000000 1.8631420 0.0000000 0.0000000 0.0000000 0.000000
#> [33,] 2.871638 0.0000000 3.7914116 0.0000000 0.0000000 0.0000000 3.046521
#> [34,] 0.000000 2.5142388 3.1312966 1.1948836 0.0000000 2.6242402 0.000000
#> [35,] 0.000000 4.0079393 0.0000000 0.0000000 2.3407402 2.0661695 0.000000
#> [36,] 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> [37,] 0.000000 0.0000000 0.0000000 2.0384512 2.1754227 0.0000000 0.000000
#> [38,]       NA 0.0000000 0.0000000 2.6198764 2.4293710 0.0000000 0.000000
#> [39,]       NA        NA 1.6092690 0.0000000 0.0000000 1.6432730 0.000000
#> [40,]       NA        NA        NA 1.3413332 0.0000000 0.0000000 0.000000
#> [41,]       NA        NA        NA        NA 1.1508969 0.0000000 3.145500
#>           [,30]     [,31]    [,32]     [,33]     [,34]     [,35]     [,36]
#>  [1,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.000000 0.0000000 0.8444607 0.0000000 0.0000000
#>  [3,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [5,] 0.0000000 0.0000000 0.000000 0.0000000 0.7214374 0.0000000 0.0000000
#>  [6,] 0.0000000 0.0000000 0.000000 0.8221505 0.0000000 0.0000000 0.0000000
#>  [7,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [8,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [9,] 0.0000000 0.0000000 0.000000 0.6270656 0.0000000 0.8640737 0.0000000
#> [10,] 0.0000000 2.0652804 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [11,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [12,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [13,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [14,] 0.0000000 0.0000000 0.000000 1.5906304 1.1243157 0.0000000 0.0000000
#> [15,] 0.0000000 0.7863596 0.000000 0.5730752 0.6104273 0.0000000 0.0000000
#> [16,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [17,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [18,] 1.0003532 1.1919918 2.470700 0.0000000 0.0000000 0.0000000 0.0000000
#> [19,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.9471291
#> [20,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.6271501 0.0000000
#> [21,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [22,] 0.0000000 0.5616322 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [23,] 0.5795834 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [24,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.5467362 0.7333738
#> [25,] 0.4509796 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [26,] 0.0000000 0.0000000 0.000000 1.1636763 0.4146091 0.5497198 0.0000000
#> [27,] 0.0000000 0.0000000 0.000000 1.4493355 0.0000000 0.0000000 0.0000000
#> [28,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [29,] 0.0000000 0.0000000 0.000000 0.0000000 1.8003397 0.0000000 0.0000000
#> [30,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [31,] 0.0000000 1.4299700 0.000000 0.0000000 0.3932143 0.0000000 0.7144827
#> [32,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [33,] 0.0000000 0.0000000 0.000000 0.0000000 1.8313124 1.2278371 2.0976768
#> [34,] 2.6577427 0.0000000 1.381844 0.5037325 1.3745707 0.0000000 0.0000000
#> [35,] 1.0111939 0.6215990 0.000000 0.9690718 0.0000000 0.0000000 0.4868882
#> [36,] 0.0000000 0.0000000 0.000000 0.0000000 0.8752836 0.0000000 1.0825044
#> [37,] 2.6537063 0.0000000 0.000000 1.9416696 0.9462451 0.0000000 0.8803134
#> [38,] 3.1180763 2.8504060 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [39,] 0.0000000 1.6322149 0.000000 0.0000000 0.0000000 0.0000000 2.7088351
#> [40,] 0.0000000 0.5199251 0.000000 1.0539764 0.0000000 0.3198205 0.0000000
#> [41,] 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           [,37]     [,38]     [,39]     [,40]     [,41]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.0000000 0.0000000 0.0000000 0.0000000 1.0033987
#>  [5,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [6,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [7,] 0.0000000 0.0000000 0.0000000 0.0000000 0.7754516
#>  [8,] 0.9557990 0.0000000 0.7792190 0.0000000 0.0000000
#>  [9,] 0.0000000 0.9762446 1.5238123 0.0000000 0.0000000
#> [10,] 0.0000000 0.0000000 0.0000000 0.8132008 0.0000000
#> [11,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [12,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [13,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [14,] 0.0000000 0.0000000 0.6589141 0.0000000 0.0000000
#> [15,] 0.0000000 0.0000000 0.0000000 0.9699457 0.0000000
#> [16,] 0.0000000 0.0000000 0.0000000 0.0000000 0.9695386
#> [17,] 0.0000000 0.9153376 0.0000000 0.0000000 0.0000000
#> [18,] 0.6649672 0.0000000 0.5589612 0.0000000 0.6073781
#> [19,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [20,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [21,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [22,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [23,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [24,] 0.5966734 1.2703551 0.0000000 0.0000000 0.0000000
#> [25,] 0.0000000 1.9950490 0.0000000 0.0000000 0.5872157
#> [26,] 0.0000000 0.0000000 1.0213538 0.0000000 0.0000000
#> [27,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [28,] 0.0000000 0.0000000 0.0000000 0.7258163 0.0000000
#> [29,] 0.0000000 1.1281071 0.0000000 1.8511341 0.0000000
#> [30,] 0.0000000 0.6754475 0.0000000 0.0000000 0.0000000
#> [31,] 0.0000000 0.6274130 0.0000000 0.0000000 0.0000000
#> [32,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [33,] 0.0000000 1.4817022 0.0000000 0.0000000 0.0000000
#> [34,] 0.4166408 0.0000000 0.0000000 0.0000000 0.0000000
#> [35,] 0.0000000 0.0000000 0.0000000 1.4491924 0.0000000
#> [36,] 0.8951552 0.0000000 0.0000000 0.0000000 0.0000000
#> [37,] 0.0000000 0.0000000 0.0000000 0.0000000 0.7325187
#> [38,] 1.9368851 1.2019191 2.4395547 0.0000000 0.0000000
#> [39,] 0.0000000 0.0000000 2.5043068 0.0000000 0.0000000
#> [40,] 0.3705595 0.0000000 0.6578825 0.0000000 0.7607654
#> [41,] 0.6688894 0.0000000 0.0000000 0.0000000 0.0000000
#> 
#> attr(,"interactions")
#> GInteractions object with 1 interaction and 15 metadata columns:
#>                 seqnames1       ranges1     seqnames2       ranges2 |
#>                     <Rle>     <IRanges>         <Rle>     <IRanges> |
#>   2L:121_2L:107        2L 120001-121000 ---        2L 106001-107000 |
#>                          name  constraint  distance orientation submatrix.name
#>                   <character> <character> <integer>   <logical>    <character>
#>   2L:121_2L:107 2L:121_2L:107       Tad_3     14000       FALSE  2L:107_2L:121
#>                  anchor.bin  anchor.name    bait.bin    bait.name
#>                 <character>  <character> <character>  <character>
#>   2L:121_2L:107      2L:121 2L:121:Tad_3      2L:107 2L:107:Tad_3
#>                 anchor.Beaf.score anchor.Beaf.name anchor.Beaf.bln
#>                         <numeric>           <list>       <numeric>
#>   2L:121_2L:107               484        Beaf32_15               1
#>                 bait.Tss.class           bait.Tss.name bait.Tss.bln
#>                    <character>                  <list>    <numeric>
#>   2L:121_2L:107         active FBgn0026787,FBgn0005278            1
#>   -------
#>   regions: 429 ranges and 0 metadata columns
#>   seqinfo: 2 sequences from an unspecified genome
#> attr(,"target")
#> attr(,"target")$anchor.Beaf.name
#> [1] "Beaf32_8"  "Beaf32_15"
#> 
#> attr(,"target")$bait.Tss.name
#> [1] "FBgn0031214" "FBgn0005278"
#> 
#> attr(,"target")$name
#> [1] "2L:74_2L:77"
#> 
#> attr(,"target")$distance
#> function(columnElement){
#>     return(14000==columnElement || columnElement == 3000)
#>     }
#> <environment: 0x55aade8b4428>
#> 
#> attr(,"selection")
#> function(){
#>   Reduce(intersect, list(anchor.Beaf.name, bait.Tss.name ,distance) ) |>
#>   setdiff(name)
#> }
#> <environment: 0x55aade8b4428>
#> attr(,"resolution")
#> [1] 1000
#> attr(,"referencePoint")
#> [1] "pf"
#> attr(,"matriceDim")
#> [1] 41

# Extraction on InteractionSet with InteractionsSet
target.lst <- list(interactions = attributes(submatrixPF_Ctrl.mtx_lst)$interactions[1:2])
FilterInteractions(
  interarctions.gni = attributes(submatrixPF_Ctrl.mtx_lst)$interactions,
  target.lst        = target.lst,
  selection.fun     = NULL
)
#> interactions1 interactions2 
#>             1             2 

# Extraction on InteractionSet list and with GRanges
gRangesTarget.gnr <- attributes(submatrixPF_Ctrl.mtx_lst)$interactions |> 
InteractionSet::anchors()
target.lst <- list(
    first = 
      gRangesTarget.gnr[["first"]][1:2]
)
FilterInteractions(
  interarctions.gni = attributes(submatrixPF_Ctrl.mtx_lst)$interactions,
  target.lst        = target.lst,
  selection.fun     = NULL
)
#> first1 first2 
#>      1      2