GetQuantif

Compute a function on each matrix that compose a matrices list.

Usage

GetQuantif(
  matrices.lst,
  area.fun = "center",
  operation.fun = "mean_rm0",
  name.chr = NULL
)

Arguments

matrices.lst

<List[matrix]> : A matrices list.

area.fun

<character or function> : A character or function thaht alow an extraction of an area on each matrix that compose the matrices lits (Default "center").

"mean_rm0" : apply a mean after replace 0 with NA

"median_rm0" : apply a median after replace 0 with NA

"sum_rm0" : apply a sum after replace 0 with NA

"median" : apply a median

"sum" : apply a sum

"mean" or other character : apply a mean

operation.fun

<character or function> : A character or function to applt on the extracted area of each matrix that compose the matrices lits (Default "mean_rm0").

"C" or "CENTER" : Take the area of interaction between anchor and bait.

"UL" or "UPPER_LEFT" : Take the area of interaction in the uppper left square

"UR" or "UPPER_RIGHT" : Take the area of interaction in the uppper right square

"BL" or "BOTTOM_LEFT" : Take the area of interaction in the bottom left square

"BR" or "BOTTOM_RIGHT" : Take the area of interaction in the bottom right square

"U" or "UPPER" : Take the area of interaction in above the center area

"B" or "BOTTOM" : Take the area of interaction in below the center area

"L" or "LEFT" : Take the area of interaction in the left of the center area

"R" or "RIGHT" : Take the area of interaction in the right of the center area

"D" or "DONUT" : Take the area of interaction that surround the center area

name.chr

<character> : The name of a column in GInteraction attributes of matrices.lst used as named in the output (Default NULL).

Value

A GRange object.