multiple GRanges Bining

Bin multiple GRanges and summary all in one GRange. Could overlap ranges with constraints regions

Usage

IndexFeatures(
  gRange.gnr_lst = NULL,
  constraint.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.num = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  cores.num = 1,
  verbose.bln = TRUE
)

Arguments

gRange.gnr_lst

<GRanges or GRangesList or list[GRanges]>: some GRanges or a list of GRanges or a GRangesList

constraint.gnr

<GRanges>: A GRange of constraint regions. If NULL chromosomes in chromSize.dtf are used (Default NULL)

chromSize.dtf

<data.frame>: A data.frame of genome where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.

binSize.num

<integer>: A number that specify the width bins.

method.chr

<character>: A string of a summary method name as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean'')

variablesName.chr_vec

<character> : A character vector that specify the metadata columns of GRanges on which apply the summary method.

cores.num

<integer> : An integer to specify the number of cores. (Default 1)

verbose.bln

<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)

Value

A GRange object.

Details

IndexFeatures

Examples

library(GenomicED)
data("anchors_Peaks.gnr")
anchors_Peaks.gnr[1]
#> GRanges object with 1 range and 2 metadata columns:
#>       seqnames      ranges strand |     score        name
#>          <Rle>   <IRanges>  <Rle> | <numeric> <character>
#>   [1]       2L 35594-35725      * |        76    Beaf32_2
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

seqlengths.num <- c('2L'=23513712, '2R'=25286936)
chromSize.dtf  <- data.frame(
  seqnames   = names(seqlengths.num ), 
  seqlengths = seqlengths.num
  )
binSize.num <- 10000

anchors_Index.gnr <- IndexFeatures(
  gRange.gnr_lst        = list(Beaf=anchors_Peaks.gnr), 
  constraint.gnr        = NULL,
  chromSize.dtf         = chromSize.dtf,
  binSize.num           = binSize.num,
  method.chr            = "max",
  variablesName.chr_vec = "score",
  cores.num             = 1,
  verbose.bln           = FALSE
  )

anchors_Index.gnr[1]
#> GRanges object with 1 range and 6 metadata columns:
#>       seqnames      ranges strand |        name         bin  constraint
#>          <Rle>   <IRanges>  <Rle> | <character> <character> <character>
#>   [1]       2L 30001-40000      * |     2L:4:2L        2L:4          2L
#>       Beaf.score Beaf.name  Beaf.bln
#>        <numeric>    <list> <numeric>
#>   [1]         76  Beaf32_2         1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome