SearchPairs

Compute pairs of genomic features that share a same range constraintes.

Usage

SearchPairs(
  indexAnchor.gnr = NULL,
  indexBait.gnr = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  verbose.bln = TRUE,
  cores.num = 1
)

Arguments

indexAnchor.gnr

<GRanges>: GRanges used in the first part of the pairs.

indexBait.gnr

<GRanges>: GRanges used in the second part of the pars. If NULL is replace with indexAnchor.gnr (Default NULL)

minDist.num

<numeric>: the minimal distance between first and second parts of pairs. (Default NULL)

maxDist.num

<numeric>: the maximal distance between first and second parts of pairs. (Default NULL)

verbose.bln

<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)

cores.num

<integer> : An integer to specify the number of cores. (Default 1)

Value

A GInteractions object.

Examples

library(GenomicED)
data(anchors_Index.gnr)
data(baits_Index.gnr)
interactions.gni <- SearchPairs(
    indexAnchor.gnr = anchors_Index.gnr,
    indexBait.gnr   = baits_Index.gnr,
    minDist.num     = 9000, 
    maxDist.num     = 11000,
    cores.num       = 1,
    verbose.bln     = FALSE
    )
interactions.gni
#> GInteractions object with 34 interactions and 15 metadata columns:
#>                     seqnames1           ranges1     seqnames2           ranges2
#>                         <Rle>         <IRanges>         <Rle>         <IRanges>
#>       2L:278_2L:268        2L     277001-278000 ---        2L     267001-268000
#>     2L:1165_2L:1174        2L   1164001-1165000 ---        2L   1173001-1174000
#>     2L:4477_2L:4466        2L   4476001-4477000 ---        2L   4465001-4466000
#>     2L:5060_2L:5071        2L   5059001-5060000 ---        2L   5070001-5071000
#>     2L:6499_2L:6509        2L   6498001-6499000 ---        2L   6508001-6509000
#>                 ...       ...               ... ...       ...               ...
#>   2R:20246_2R:20255        2R 20245001-20246000 ---        2R 20254001-20255000
#>   2R:20668_2R:20658        2R 20667001-20668000 ---        2R 20657001-20658000
#>   2R:20669_2R:20658        2R 20668001-20669000 ---        2R 20657001-20658000
#>   2R:22103_2R:22114        2R 22102001-22103000 ---        2R 22113001-22114000
#>   2R:22104_2R:22114        2R 22103001-22104000 ---        2R 22113001-22114000
#>                     |              name  constraint  distance orientation
#>                     |       <character> <character> <integer>   <logical>
#>       2L:278_2L:268 |     2L:278_2L:268      Tad_10     10000       FALSE
#>     2L:1165_2L:1174 |   2L:1165_2L:1174      Tad_35      9000        TRUE
#>     2L:4477_2L:4466 |   2L:4477_2L:4466     Tad_114     11000       FALSE
#>     2L:5060_2L:5071 |   2L:5060_2L:5071     Tad_132     11000        TRUE
#>     2L:6499_2L:6509 |   2L:6499_2L:6509     Tad_162     10000        TRUE
#>                 ... .               ...         ...       ...         ...
#>   2R:20246_2R:20255 | 2R:20246_2R:20255     Tad_905      9000        TRUE
#>   2R:20668_2R:20658 | 2R:20668_2R:20658     Tad_917     10000       FALSE
#>   2R:20669_2R:20658 | 2R:20669_2R:20658     Tad_917     11000       FALSE
#>   2R:22103_2R:22114 | 2R:22103_2R:22114     Tad_957     11000        TRUE
#>   2R:22104_2R:22114 | 2R:22104_2R:22114     Tad_957     10000        TRUE
#>                        submatrix.name  anchor.bin      anchor.name    bait.bin
#>                           <character> <character>      <character> <character>
#>       2L:278_2L:268     2L:268_2L:278      2L:278    2L:278:Tad_10      2L:268
#>     2L:1165_2L:1174   2L:1165_2L:1174     2L:1165   2L:1165:Tad_35     2L:1174
#>     2L:4477_2L:4466   2L:4466_2L:4477     2L:4477  2L:4477:Tad_114     2L:4466
#>     2L:5060_2L:5071   2L:5060_2L:5071     2L:5060  2L:5060:Tad_132     2L:5071
#>     2L:6499_2L:6509   2L:6499_2L:6509     2L:6499  2L:6499:Tad_162     2L:6509
#>                 ...               ...         ...              ...         ...
#>   2R:20246_2R:20255 2R:20246_2R:20255    2R:20246 2R:20246:Tad_905    2R:20255
#>   2R:20668_2R:20658 2R:20658_2R:20668    2R:20668 2R:20668:Tad_917    2R:20658
#>   2R:20669_2R:20658 2R:20658_2R:20669    2R:20669 2R:20669:Tad_917    2R:20658
#>   2R:22103_2R:22114 2R:22103_2R:22114    2R:22103 2R:22103:Tad_957    2R:22114
#>   2R:22104_2R:22114 2R:22104_2R:22114    2R:22104 2R:22104:Tad_957    2R:22114
#>                            bait.name anchor.Beaf.score        anchor.Beaf.name
#>                          <character>         <numeric>                  <list>
#>       2L:278_2L:268    2L:268:Tad_10              1641               Beaf32_41
#>     2L:1165_2L:1174   2L:1174:Tad_35               113              Beaf32_127
#>     2L:4477_2L:4466  2L:4466:Tad_114                54              Beaf32_341
#>     2L:5060_2L:5071  2L:5071:Tad_132               433              Beaf32_375
#>     2L:6499_2L:6509  2L:6509:Tad_162               293              Beaf32_461
#>                 ...              ...               ...                     ...
#>   2R:20246_2R:20255 2R:20255:Tad_905              1933             Beaf32_2614
#>   2R:20668_2R:20658 2R:20658:Tad_917              1422 Beaf32_2633,Beaf32_2634
#>   2R:20669_2R:20658 2R:20658:Tad_917              1422             Beaf32_2634
#>   2R:22103_2R:22114 2R:22114:Tad_957               521             Beaf32_2725
#>   2R:22104_2R:22114 2R:22114:Tad_957               521             Beaf32_2725
#>                     anchor.Beaf.bln bait.Tss.class bait.Tss.name bait.Tss.bln
#>                           <numeric>    <character>        <list>    <numeric>
#>       2L:278_2L:268               1       inactive   FBgn0031239            1
#>     2L:1165_2L:1174               1         active   FBgn0031318            1
#>     2L:4477_2L:4466               1         active   FBgn0031611            1
#>     2L:5060_2L:5071               1         active   FBgn0261608            1
#>     2L:6499_2L:6509               1       inactive   FBgn0266001            1
#>                 ...             ...            ...           ...          ...
#>   2R:20246_2R:20255               1       inactive   FBgn0034490            1
#>   2R:20668_2R:20658               1       inactive   FBgn0050155            1
#>   2R:20669_2R:20658               1       inactive   FBgn0050155            1
#>   2R:22103_2R:22114               1       inactive   FBgn0050280            1
#>   2R:22104_2R:22114               1       inactive   FBgn0050280            1
#>   -------
#>   regions: 429 ranges and 0 metadata columns
#>   seqinfo: 2 sequences from an unspecified genome