Aggregation of matrices list.

Aggregates all the matrices of a list (or two lists in case of differential aggregation) into a single matrix. This function allows to apply different aggregation (average, sum, ...), transformation (rank, percentage, ...) and differential (subtraction, ratio, ...) functions.

Usage

Aggregation(
  ctrlMatrices.lst = NULL,
  matrices.lst = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  trans.fun = NULL,
  agg.fun = "mean",
  rm0.bln = FALSE,
  diff.fun = "substraction",
  scaleCorrection.bln = FALSE,
  correctionArea.lst = NULL,
  statCompare.bln = FALSE
)

Arguments

ctrlMatrices.lst

<list[matrix]>: The matrices list to aggregate as control.

matrices.lst

<list[matrix]>: The matrices list to aggregate.

minDist.num

<numeric>: The minimal distance between anchor and bait.

maxDist.num

<numeric>: The maximal distance between anchor and bait.

trans.fun

<function or chracter>: The function use to transforme or scale values in each submatrix before aggregation. If the parameter is character so:

"quantile" or "qtl" apply function dplyr::ntile(x,500)

"percentile" or "prct" apply percentile.

"rank" apply a ranking.

"zscore" apply a scaling.

"minmax" apply a Genomic2DTK::MinMaxScale.

"mu" apply a Genomic2DTK::MeanScale.

other or NULL don't apply transformation (Default).

agg.fun

<function or chracter>: The function use to aggregate each pixel in matrix. If the parameter is a character so:

"50

"+" or "sum" apply the sum

other (Default) apply the mean

rm0.bln

<logical>: if TURE 0 are replace with NA. (Default FALSE)

diff.fun

<function or chracter>: The function use to compute differential. If the parameter is character so:

"-", "substract" or "substraction" apply a substraction

"/" or "ratio" apply a ratio

"log2","log2-","log2/" or "log2ratio" apply a log2 on ratio

other (Default) apply a log2 on 1+ratio

scaleCorrection.bln

<logical>: Whether a correction should be done on the median value take in ane noising area. (Default TRUE)

correctionArea.lst

<list>: Nested list of indice that define a noising area fore correction. List muste contain in first an element "i" (row indices) then an element called "j" (columns indices). If NULL automatically take in upper left part of aggregated matrices. (Default NULL)

statCompare.bln

<logical>: Whether a t.test must be apply to each pxl of the differential aggregated matrix.

Value

A matrix

Details

Aggregation

Examples

if (FALSE) {
  # Classical Aggregation
    # submatrixRF_Ctrl.mtx_lst is a submatrix list
    # obtain with Genomic2DTK::ExtractSubmatrix function
  aggreg.mtx <- Aggregation(
    matrices.lst = submatrixRF_Ctrl.mtx_lst, 
    agg.fun      = "sum",
    trans.fun    = "qtl", 
    rm0.bln      = TRUE,
    minDist      = 9000,
    maxDist      = 11000
  )
 
  # Differential Aggregation
    # submatrixRF.mtx_lst is a second submatrix list
    # obtain with Genomic2DTK::ExtractSubmatrix function
  diffAggreg.mtx <- Aggregation(
    ctrlMatrices.lst    = submatrixRF_Ctrl.mtx_lst,
    matrices.lst        = submatrixRF.mtx_lst,
    minDist             = 9000,
    maxDist             = 11000,
    agg.fun             = "mean",
    rm0.bln             = FALSE,
    diff.fun            = "substraction",
    scaleCorrection.bln = TRUE,
    correctionArea.lst  =  list(
      i = c(1:30),
      j = c(72:101)
      ),
    statCompare.bln = TRUE
  )
}