GetQuantif

Compute a function on each matrix that compose a matrices list.

Usage

GetQuantif(
  matrices.lst,
  area.fun = "center",
  operation.fun = "mean_rm0",
  name.chr = NULL
)

Arguments

matrices.lst

<List[matrix]> : A matrices list.

area.fun

<character or function> : A character or function thaht alow an extraction of an area on each matrix that compose the matrices lits (Default "center").

"mean_rm0" : apply a mean after replace 0 with NA

"median_rm0" : apply a median after replace 0 with NA

"sum_rm0" : apply a sum after replace 0 with NA

"median" : apply a median

"sum" : apply a sum

"mean" or other character : apply a mean

operation.fun

<character or function> : A character or function to applt on the extracted area of each matrix that compose the matrices lits (Default "mean_rm0").

"C" or "CENTER" : Take the area of interaction between anchor and bait.

"UL" or "UPPER_LEFT" : Take the area of interaction in the uppper left square

"UR" or "UPPER_RIGHT" : Take the area of interaction in the uppper right square

"BL" or "BOTTOM_LEFT" : Take the area of interaction in the bottom left square

"BR" or "BOTTOM_RIGHT" : Take the area of interaction in the bottom right square

"U" or "UPPER" : Take the area of interaction in above the center area

"B" or "BOTTOM" : Take the area of interaction in below the center area

"L" or "LEFT" : Take the area of interaction in the left of the center area

"R" or "RIGHT" : Take the area of interaction in the right of the center area

"D" or "DONUT" : Take the area of interaction that surround the center area

name.chr

<character> : The name of a column in GInteraction attributes of matrices.lst used as named in the output (Default NULL).

Value

A GRange object.

Examples

library(GenomicED)
data("submatrixRF_Ctrl.mtx_lst")
GetQuantif(
  matrices.lst  = submatrixRF_Ctrl.mtx_lst,
  area.fun      = "center",
  operation.fun = "mean"
  )
#> $`2L:74_2L:77`
#> [1] 0.002752447
#> 
#> $`2L:107_2L:115`
#> [1] 0.004792452
#> 
#> $`2L:107_2L:121`
#> [1] 1.01102
#> 
#> $`2L:117_2L:121`
#> [1] 0.03643152
#> 
#> $`2L:268_2L:274`
#> [1] 1.69169
#> 
#> $`2L:268_2L:278`
#> [1] 1.483054
#> 
#> $`2L:274_2L:279`
#> [1] 1.699314
#> 
#> $`2L:306_2L:320`
#> [1] 0.6724637
#> 
#> $`2L:422_2L:471`
#> [1] 0.3981305
#> 
#> $`2L:422_2L:472`
#> [1] 0.2917365
#> 
#> $`2L:454_2L:471`
#> [1] 0.4176653
#> 
#> $`2L:454_2L:472`
#> [1] 0.07697316
#> 
#> $`2L:1165_2L:1174`
#> [1] 0.1933935
#> 
#> $`2L:1174_2L:1178`
#> [1] 1.97396
#> 
#> $`2L:2493_2L:2534`
#> [1] 0.1509465
#> 
#> $`2L:2494_2L:2534`
#> [1] 0.1147519
#> 
#> $`2L:2654_2L:2678`
#> [1] 0.03862693
#> 
#> $`2L:2771_2L:2805`
#> [1] 0.1995228
#> 
#> $`2L:2771_2L:2806`
#> [1] 0.1731242
#> 
#> $`2L:2980_2L:2993`
#> [1] 1.403238
#> 
#> $`2L:2988_2L:2993`
#> [1] 1.966797
#> 
#> $`2L:2989_2L:2993`
#> [1] 2.430412
#> 
#> $`2L:3160_2L:3164`
#> [1] 0
#> 
#> $`2L:3333_2L:3347`
#> [1] 0.04267804
#> 
#> $`2L:3333_2L:3349`
#> [1] 0.8697373
#> 
#> $`2L:3333_2L:3355`
#> [1] 0.1149894
#> 
#> $`2L:3718_2L:3731`
#> [1] 0.005324978
#> 
#> $`2L:4404_2L:4434`
#> [1] 0.706205
#> 
#> $`2L:4466_2L:4477`
#> [1] 0.8661045
#> 
#> $`2L:4466_2L:4478`
#> [1] 0.08887885
#> 
#> $`2L:4980_2L:4999`
#> [1] 0.5829616
#> 
#> $`2L:5052_2L:5059`
#> [1] 0.01361992
#> 
#> $`2L:5059_2L:5071`
#> [1] 0.07153248
#> 
#> $`2L:5060_2L:5071`
#> [1] 0
#> 
#> $`2L:5071_2L:5094`
#> [1] 0.2870154
#> 
#> $`2L:5521_2L:5529`
#> [1] 0.08567937
#> 
#> $`2L:5521_2L:5533`
#> [1] 0.07505932
#> 
#> $`2L:5526_2L:5529`
#> [1] 0.07977823
#> 
#> $`2L:5526_2L:5533`
#> [1] 0.03038824
#> 
#> $`2L:5527_2L:5533`
#> [1] 0.5756298
#> 
#> $`2L:5566_2L:5725`
#> [1] 0.6408698
#> 
#> $`2L:5616_2L:5725`
#> [1] 0.1162785
#> 
#> $`2L:5633_2L:5725`
#> [1] 0.3059426
#> 
#> $`2L:5908_2L:5941`
#> [1] 0.3542089
#> 
#> $`2L:5909_2L:5941`
#> [1] 0.1871565
#> 
#> $`2L:5944_2L:5968`
#> [1] 0
#> 
#> $`2L:5945_2L:5968`
#> [1] 0
#> 
#> $`2L:6051_2L:6069`
#> [1] 0.5279512
#> 
#> $`2L:6066_2L:6069`
#> [1] 2.139299
#> 
#> $`2L:6071_2L:6084`
#> [1] 0.1096642
#> 
#> $`2L:6071_2L:6089`
#> [1] 0.1007996
#> 
#> $`2L:6499_2L:6509`
#> [1] 1.630512
#> 
#> $`2L:6724_2L:6743`
#> [1] 0.1687783
#> 
#> $`2L:6908_2L:6915`
#> [1] 0.134238
#> 
#> $`2L:6908_2L:6919`
#> [1] 0.08672329
#> 
#> $`2L:7033_2L:7038`
#> [1] 1.703114
#> 
#> $`2L:7033_2L:7039`
#> [1] 1.601052
#> 
#> $`2L:7066_2L:7085`
#> [1] 0.4551538
#> 
#> $`2L:7066_2L:7156`
#> [1] 0.09590472
#> 
#> $`2L:7085_2L:7179`
#> [1] 0.9786903
#> 
#> $`2L:7156_2L:7179`
#> [1] 0.4174838
#> 
#> $`2L:7812_2L:7828`
#> [1] 0.8693823
#> 
#> $`2L:8125_2L:8128`
#> [1] 0.06835569
#> 
#> $`2L:8242_2L:8279`
#> [1] 0.6649474
#> 
#> $`2L:8240_2L:8243`
#> [1] 0.002060519
#> 
#> $`2L:8243_2L:8279`
#> [1] 0.8386224
#> 
#> $`2L:8687_2L:8710`
#> [1] 0.1404249
#> 
#> $`2L:8694_2L:8710`
#> [1] 0.04914559
#> 
#> $`2L:9760_2L:9764`
#> [1] 0.912675
#> 
#> $`2L:9764_2L:9780`
#> [1] 0.1105858
#> 
#> $`2L:9764_2L:9783`
#> [1] 0.1230102
#> 
#> $`2L:9760_2L:9776`
#> [1] 0.6786809
#> 
#> $`2L:9776_2L:9780`
#> [1] 0.008095992
#> 
#> $`2L:9776_2L:9783`
#> [1] 0.08259367
#> 
#> $`2L:9985_2L:10003`
#> [1] 0.2712008
#> 
#> $`2L:10233_2L:10248`
#> [1] 0.02951273
#> 
#> $`2L:10387_2L:10394`
#> [1] 0.1361494
#> 
#> $`2L:10388_2L:10394`
#> [1] 0.09456354
#> 
#> $`2L:10389_2L:10394`
#> [1] 0.009158009
#> 
#> $`2L:10394_2L:10397`
#> [1] 0.1372137
#> 
#> $`2L:10696_2L:10729`
#> [1] 0
#> 
#> $`2L:10702_2L:10729`
#> [1] 0.1690664
#> 
#> $`2L:10729_2L:10738`
#> [1] 0.02811435
#> 
#> $`2L:10730_2L:10738`
#> [1] 0
#> 
#> $`2L:11949_2L:11989`
#> [1] 0.07958475
#> 
#> $`2L:11989_2L:12006`
#> [1] 0.2831623
#> 
#> $`2L:11949_2L:12000`
#> [1] 0.004800573
#> 
#> $`2L:11990_2L:12000`
#> [1] 0.03047268
#> 
#> $`2L:11991_2L:12000`
#> [1] 0.3249912
#> 
#> $`2L:12000_2L:12006`
#> [1] 1.986563
#> 
#> $`2L:11949_2L:12026`
#> [1] 0
#> 
#> $`2L:11990_2L:12026`
#> [1] 0.05402907
#> 
#> $`2L:11991_2L:12026`
#> [1] 0.2037945
#> 
#> $`2L:12006_2L:12026`
#> [1] 0.3690407
#> 
#> $`2L:12431_2L:12435`
#> [1] 0.8919457
#> 
#> $`2L:12498_2L:12515`
#> [1] 0
#> 
#> $`2L:12498_2L:12516`
#> [1] 0
#> 
#> $`2L:12508_2L:12515`
#> [1] 0.4756325
#> 
#> $`2L:12508_2L:12516`
#> [1] 0.02889891
#> 
#> $`2L:12509_2L:12515`
#> [1] 0.1061493
#> 
#> attr(,"interactions")
#> GInteractions object with 100 interactions and 15 metadata columns:
#>                     seqnames1           ranges1     seqnames2           ranges2
#>                         <Rle>         <IRanges>         <Rle>         <IRanges>
#>         2L:74_2L:77        2L       73001-74000 ---        2L       76001-77000
#>       2L:115_2L:107        2L     114001-115000 ---        2L     106001-107000
#>       2L:121_2L:107        2L     120001-121000 ---        2L     106001-107000
#>       2L:121_2L:117        2L     120001-121000 ---        2L     116001-117000
#>       2L:274_2L:268        2L     273001-274000 ---        2L     267001-268000
#>                 ...       ...               ... ...       ...               ...
#>   2L:12515_2L:12498        2L 12514001-12515000 ---        2L 12497001-12498000
#>   2L:12516_2L:12498        2L 12515001-12516000 ---        2L 12497001-12498000
#>   2L:12515_2L:12508        2L 12514001-12515000 ---        2L 12507001-12508000
#>   2L:12516_2L:12508        2L 12515001-12516000 ---        2L 12507001-12508000
#>   2L:12515_2L:12509        2L 12514001-12515000 ---        2L 12508001-12509000
#>                     |              name  constraint  distance orientation
#>                     |       <character> <character> <integer>   <logical>
#>         2L:74_2L:77 |       2L:74_2L:77       Tad_1      3000        TRUE
#>       2L:115_2L:107 |     2L:115_2L:107       Tad_3      8000       FALSE
#>       2L:121_2L:107 |     2L:121_2L:107       Tad_3     14000       FALSE
#>       2L:121_2L:117 |     2L:121_2L:117       Tad_3      4000       FALSE
#>       2L:274_2L:268 |     2L:274_2L:268      Tad_10      6000       FALSE
#>                 ... .               ...         ...       ...         ...
#>   2L:12515_2L:12498 | 2L:12515_2L:12498     Tad_310     17000       FALSE
#>   2L:12516_2L:12498 | 2L:12516_2L:12498     Tad_310     18000       FALSE
#>   2L:12515_2L:12508 | 2L:12515_2L:12508     Tad_310      7000       FALSE
#>   2L:12516_2L:12508 | 2L:12516_2L:12508     Tad_310      8000       FALSE
#>   2L:12515_2L:12509 | 2L:12515_2L:12509     Tad_310      6000       FALSE
#>                        submatrix.name  anchor.bin      anchor.name    bait.bin
#>                           <character> <character>      <character> <character>
#>         2L:74_2L:77       2L:74_2L:77       2L:74      2L:74:Tad_1       2L:77
#>       2L:115_2L:107     2L:107_2L:115      2L:115     2L:115:Tad_3      2L:107
#>       2L:121_2L:107     2L:107_2L:121      2L:121     2L:121:Tad_3      2L:107
#>       2L:121_2L:117     2L:117_2L:121      2L:121     2L:121:Tad_3      2L:117
#>       2L:274_2L:268     2L:268_2L:274      2L:274    2L:274:Tad_10      2L:268
#>                 ...               ...         ...              ...         ...
#>   2L:12515_2L:12498 2L:12498_2L:12515    2L:12515 2L:12515:Tad_310    2L:12498
#>   2L:12516_2L:12498 2L:12498_2L:12516    2L:12516 2L:12516:Tad_310    2L:12498
#>   2L:12515_2L:12508 2L:12508_2L:12515    2L:12515 2L:12515:Tad_310    2L:12508
#>   2L:12516_2L:12508 2L:12508_2L:12516    2L:12516 2L:12516:Tad_310    2L:12508
#>   2L:12515_2L:12509 2L:12509_2L:12515    2L:12515 2L:12515:Tad_310    2L:12509
#>                            bait.name anchor.Beaf.score anchor.Beaf.name
#>                          <character>         <numeric>           <list>
#>         2L:74_2L:77      2L:77:Tad_1               205         Beaf32_8
#>       2L:115_2L:107     2L:107:Tad_3              1707        Beaf32_14
#>       2L:121_2L:107     2L:107:Tad_3               484        Beaf32_15
#>       2L:121_2L:117     2L:117:Tad_3               484        Beaf32_15
#>       2L:274_2L:268    2L:268:Tad_10               210        Beaf32_39
#>                 ...              ...               ...              ...
#>   2L:12515_2L:12498 2L:12498:Tad_310                44       Beaf32_900
#>   2L:12516_2L:12498 2L:12498:Tad_310                88       Beaf32_901
#>   2L:12515_2L:12508 2L:12508:Tad_310                44       Beaf32_900
#>   2L:12516_2L:12508 2L:12508:Tad_310                88       Beaf32_901
#>   2L:12515_2L:12509 2L:12509:Tad_310                44       Beaf32_900
#>                     anchor.Beaf.bln bait.Tss.class           bait.Tss.name
#>                           <numeric>    <character>                  <list>
#>         2L:74_2L:77               1       inactive             FBgn0031214
#>       2L:115_2L:107               1         active FBgn0026787,FBgn0005278
#>       2L:121_2L:107               1         active FBgn0026787,FBgn0005278
#>       2L:121_2L:117               1         active             FBgn0031219
#>       2L:274_2L:268               1       inactive             FBgn0031239
#>                 ...             ...            ...                     ...
#>   2L:12515_2L:12498               1       inactive             FBgn0051858
#>   2L:12516_2L:12498               1       inactive             FBgn0051858
#>   2L:12515_2L:12508               1       inactive             FBgn0262353
#>   2L:12516_2L:12508               1       inactive             FBgn0262353
#>   2L:12515_2L:12509               1       inactive             FBgn0263458
#>                     bait.Tss.bln
#>                        <numeric>
#>         2L:74_2L:77            1
#>       2L:115_2L:107            1
#>       2L:121_2L:107            1
#>       2L:121_2L:117            1
#>       2L:274_2L:268            1
#>                 ...          ...
#>   2L:12515_2L:12498            1
#>   2L:12516_2L:12498            1
#>   2L:12515_2L:12508            1
#>   2L:12516_2L:12508            1
#>   2L:12515_2L:12509            1
#>   -------
#>   regions: 429 ranges and 0 metadata columns
#>   seqinfo: 2 sequences from an unspecified genome
#> attr(,"target")
#> attr(,"target")$submatrix.name
#>   [1] "2L:74_2L:77"       "2L:107_2L:115"     "2L:107_2L:121"    
#>   [4] "2L:117_2L:121"     "2L:268_2L:274"     "2L:268_2L:278"    
#>   [7] "2L:274_2L:279"     "2L:306_2L:320"     "2L:422_2L:471"    
#>  [10] "2L:422_2L:472"     "2L:454_2L:471"     "2L:454_2L:472"    
#>  [13] "2L:1165_2L:1174"   "2L:1174_2L:1178"   "2L:2493_2L:2534"  
#>  [16] "2L:2494_2L:2534"   "2L:2654_2L:2678"   "2L:2771_2L:2805"  
#>  [19] "2L:2771_2L:2806"   "2L:2980_2L:2993"   "2L:2988_2L:2993"  
#>  [22] "2L:2989_2L:2993"   "2L:3160_2L:3164"   "2L:3333_2L:3347"  
#>  [25] "2L:3333_2L:3349"   "2L:3333_2L:3355"   "2L:3718_2L:3731"  
#>  [28] "2L:4404_2L:4434"   "2L:4466_2L:4477"   "2L:4466_2L:4478"  
#>  [31] "2L:4980_2L:4999"   "2L:5052_2L:5059"   "2L:5059_2L:5071"  
#>  [34] "2L:5060_2L:5071"   "2L:5071_2L:5094"   "2L:5521_2L:5529"  
#>  [37] "2L:5521_2L:5533"   "2L:5526_2L:5529"   "2L:5526_2L:5533"  
#>  [40] "2L:5527_2L:5533"   "2L:5566_2L:5725"   "2L:5616_2L:5725"  
#>  [43] "2L:5633_2L:5725"   "2L:5908_2L:5941"   "2L:5909_2L:5941"  
#>  [46] "2L:5944_2L:5968"   "2L:5945_2L:5968"   "2L:6051_2L:6069"  
#>  [49] "2L:6066_2L:6069"   "2L:6071_2L:6084"   "2L:6071_2L:6089"  
#>  [52] "2L:6499_2L:6509"   "2L:6724_2L:6743"   "2L:6908_2L:6915"  
#>  [55] "2L:6908_2L:6919"   "2L:7033_2L:7038"   "2L:7033_2L:7039"  
#>  [58] "2L:7066_2L:7085"   "2L:7066_2L:7156"   "2L:7085_2L:7179"  
#>  [61] "2L:7156_2L:7179"   "2L:7812_2L:7828"   "2L:8125_2L:8128"  
#>  [64] "2L:8242_2L:8279"   "2L:8240_2L:8243"   "2L:8243_2L:8279"  
#>  [67] "2L:8687_2L:8710"   "2L:8694_2L:8710"   "2L:9760_2L:9764"  
#>  [70] "2L:9764_2L:9780"   "2L:9764_2L:9783"   "2L:9760_2L:9776"  
#>  [73] "2L:9776_2L:9780"   "2L:9776_2L:9783"   "2L:9985_2L:10003" 
#>  [76] "2L:10233_2L:10248" "2L:10387_2L:10394" "2L:10388_2L:10394"
#>  [79] "2L:10389_2L:10394" "2L:10394_2L:10397" "2L:10696_2L:10729"
#>  [82] "2L:10702_2L:10729" "2L:10729_2L:10738" "2L:10730_2L:10738"
#>  [85] "2L:11949_2L:11989" "2L:11989_2L:12006" "2L:11949_2L:12000"
#>  [88] "2L:11990_2L:12000" "2L:11991_2L:12000" "2L:12000_2L:12006"
#>  [91] "2L:11949_2L:12026" "2L:11990_2L:12026" "2L:11991_2L:12026"
#>  [94] "2L:12006_2L:12026" "2L:12431_2L:12435" "2L:12498_2L:12515"
#>  [97] "2L:12498_2L:12516" "2L:12508_2L:12515" "2L:12508_2L:12516"
#> [100] "2L:12509_2L:12515"
#> 
#> attr(,"resolution")
#> [1] 1000
#> attr(,"referencePoint")
#> [1] "rf"
#> attr(,"matriceDim")
#> [1] 101
#> attr(,"shiftFactor")
#> [1] 1
#> attr(,"operation")
#> function (x) 
#> {
#>     mean(x, na.rm = TRUE)
#> }
#> <environment: 0x55a18278d938>
#> attr(,"area")
#> function (matrice.mtx) 
#> {
#>     matrice.mtx[33:35, 67:69]
#> }
#> <environment: 0x55a17fb91e60>