Creates pairs of coordinates from indexed anchor and bait genomic coordinates according to distance constraints.

Creates pairs of coordinates from indexed anchor and bait genomic coordinates according to distance constraints.

Usage

SearchPairs(
  indexAnchor.gnr = NULL,
  indexBait.gnr = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  verbose.bln = TRUE,
  cores.num = 1
)

Arguments

indexAnchor.gnr

<GRanges>: A first indexed GRanges object used as pairs anchor (must be indexed using IndexFeatures()).

indexBait.gnr

<GRanges>: A second indexed GRanges object used as pairs bait (must be indexed using IndexFeatures()). If NULL, indexAnchor.gnr is used instead (Default NULL)

minDist.num

<numeric>: Minimal distance between anchors and baits. (Default NULL)

maxDist.num

<numeric>: Maximal distance between anchors and baits. (Default NULL)

verbose.bln

<logical>: If TRUE show the progression in console. (Default TRUE)

cores.num

<integer> : Number of cores to use. (Default 1)

Value

A GInteractions object.

Details

SearchPairs

Examples

library(GenomicED)

data(anchors_Index.gnr)
anchors_Index.gnr[1]
#> GRanges object with 1 range and 6 metadata columns:
#>       seqnames      ranges strand |        name         bin  constraint
#>          <Rle>   <IRanges>  <Rle> | <character> <character> <character>
#>   [1]       2L 73001-74000      * | 2L:74:Tad_1       2L:74       Tad_1
#>       Beaf.score Beaf.name  Beaf.bln
#>        <numeric>    <list> <numeric>
#>   [1]        205  Beaf32_8         1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome

data(baits_Index.gnr)
baits_Index.gnr[1]
#> GRanges object with 1 range and 6 metadata columns:
#>       seqnames      ranges strand |        name         bin  constraint
#>          <Rle>   <IRanges>  <Rle> | <character> <character> <character>
#>   [1]       2L 76001-77000      * | 2L:77:Tad_1       2L:77       Tad_1
#>         Tss.class    Tss.name   Tss.bln
#>       <character>      <list> <numeric>
#>   [1]    inactive FBgn0031214         1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome

interactions.gni <- SearchPairs(
    indexAnchor.gnr = anchors_Index.gnr,
    indexBait.gnr   = baits_Index.gnr,
    minDist.num     = 9000, 
    maxDist.num     = 11000,
    cores.num       = 1,
    verbose.bln     = FALSE
    )

interactions.gni[1]
#> GInteractions object with 1 interaction and 15 metadata columns:
#>                 seqnames1       ranges1     seqnames2       ranges2 |
#>                     <Rle>     <IRanges>         <Rle>     <IRanges> |
#>   2L:278_2L:268        2L 277001-278000 ---        2L 267001-268000 |
#>                          name  constraint  distance orientation submatrix.name
#>                   <character> <character> <integer>   <logical>    <character>
#>   2L:278_2L:268 2L:278_2L:268      Tad_10     10000       FALSE  2L:268_2L:278
#>                  anchor.bin   anchor.name    bait.bin     bait.name
#>                 <character>   <character> <character>   <character>
#>   2L:278_2L:268      2L:278 2L:278:Tad_10      2L:268 2L:268:Tad_10
#>                 anchor.Beaf.score anchor.Beaf.name anchor.Beaf.bln
#>                         <numeric>           <list>       <numeric>
#>   2L:278_2L:268              1641        Beaf32_41               1
#>                 bait.Tss.class bait.Tss.name bait.Tss.bln
#>                    <character>        <list>    <numeric>
#>   2L:278_2L:268       inactive   FBgn0031239            1
#>   -------
#>   regions: 429 ranges and 0 metadata columns
#>   seqinfo: 2 sequences from an unspecified genome